+++ src/Stub/Answer/CallRequest.php
@@ -13,6 +13,7 @@ namespace Eloquent\Phony\Stub\Answer;
 
 use Eloquent\Phony\Call\Argument\Arguments;
 use Eloquent\Phony\Call\Argument\ArgumentsInterface;
+use Eloquent\Phony\Mock\Proxy\InstanceProxyInterface;
 
 /**
  * Represents a call request.
@@ -40,6 +41,15 @@ class CallRequest implements CallRequestInterface
         $this->prefixSelf = $prefixSelf;
         $this->suffixArgumentsObject = $suffixArgumentsObject;
         $this->suffixArguments = $suffixArguments;
+
+        foreach ($this->arguments->all() as $index => $argument) {
+            if (
+                $argument instanceof InstanceProxyInterface &&
+                $argument->isAdaptable()
+            ) {
+                $this->arguments->set($index, $argument->mock());
+            }
+        }
     }
 
     /**